sched: fair: fix long-term EEVDF drift
- place_entity: remove max_vruntime clamp that was erasing positive
  lag on every wakeup. Tasks that were owed CPU time got placed at
  avg_vruntime instead of ahead of it, causing stuttering after
  hours of use.
- update_curr: reorder update_deadline before update_min_vruntime
  to match mainline 6.6 ordering.
- put_prev_entity: remove extra update_entity_lag call that mainline
  6.6 doesn't have. Lag is only computed in dequeue_entity.
- sched.h: change load_sum from u32 to long to match mainline 6.6
  and prevent potential truncation in avg_vruntime_update.

+++ kernel/sched/sched.h
@@ -631,7 +631,7 @@ struct cfs_rq {
 
 #ifdef CONFIG_SCHED_EEVDF
 	s64			weighted_vruntime_sum;
-	u32			load_sum;
+	long			load_sum;
 #endif
 
 	struct rb_root_cached	tasks_timeline;
